DKO NA LAAM

+++ additionprac.zig
@@ -1,8 +1,8 @@
 const std = @import("std");
 
 pub fn main() void {
-    const a: i32 = 500;
-    const b: i32 = 400;
+    const a: f64 = 500;
+    const b: f64 = 400;
 
     const result = a + b;
